+++ DemoApp/Screens/UserProfileViewController.swift
@@ -5,10 +5,18 @@
 import StreamChat
 import UIKit
 
-class UserProfileViewController: UIViewController, CurrentChatUserControllerDelegate {
+class UserProfileViewController: UITableViewController, CurrentChatUserControllerDelegate {
     private let imageView = UIImageView()
-    private let nameTextField = UITextField()
     private let updateButton = UIButton()
+
+    var name: String?
+    let properties = UserProperty.allCases
+
+    enum UserProperty: CaseIterable {
+        case name
+        case typingIndicatorsEnabled
+        case readReceiptsEnabled
+    }
 
     let currentUserController: CurrentChatUserController
 
@@ -25,12 +33,17 @@ class UserProfileViewController: UIViewController, CurrentChatUserControllerDele
     override func viewDidLoad() {
         super.viewDidLoad()
 
+        tableView.allowsSelection = false
         view.backgroundColor = .systemBackground
 
-        [imageView, nameTextField, updateButton].forEach {
-            view.addSubview($0)
+        [imageView, updateButton].forEach {
             $0.translatesAutoresizingMaskIntoConstraints = false
         }
+
+        tableView.tableHeaderView = UIView(frame: .init(origin: .zero, size: .init(width: .zero, height: 80)))
+        tableView.tableHeaderView?.addSubview(imageView)
+        tableView.tableFooterView = UIView(frame: .init(origin: .zero, size: .init(width: .zero, height: 80)))
+        tableView.tableFooterView?.addSubview(updateButton)
 
         imageView.contentMode = .scaleAspectFill
         imageView.layer.cornerRadius = 30
@@ -46,45 +59,90 @@ class UserProfileViewController: UIViewController, CurrentChatUserControllerDele
             imageView.widthAnchor.constraint(equalToConstant: 60),
             imageView.heightAnchor.constraint(equalTo: imageView.widthAnchor),
             imageView.centerXAnchor.constraint(equalTo: view.centerXAnchor),
-            imageView.topAnchor.constraint(equalTo: view.safeAreaLayoutGuide.topAnchor, constant: 20),
-            nameTextField.centerXAnchor.constraint(equalTo: view.centerXAnchor),
-            nameTextField.topAnchor.constraint(equalTo: imageView.bottomAnchor, constant: 20),
             updateButton.centerXAnchor.constraint(equalTo: view.centerXAnchor),
-            updateButton.topAnchor.constraint(equalTo: nameTextField.bottomAnchor, constant: 20),
-            updateButton.heightAnchor.constraint(equalToConstant: 35)
+            updateButton.heightAnchor.constraint(equalToConstant: 35),
+            updateButton.centerYAnchor.constraint(equalTo: updateButton.superview!.centerYAnchor)
         ])
 
         currentUserController.delegate = self
         synchronizeAndUpdateData()
     }
 
-    private func synchronizeAndUpdateData() {
-        currentUserController.synchronize { [weak self] _ in
-            self?.updateUserData()
+    override func tableView(_ tableView: UITableView, numberOfRowsInSection section: Int) -> Int {
+        properties.count
+    }
+
+    override func tableView(_ tableView: UITableView, cellForRowAt indexPath: IndexPath) -> UITableViewCell {
+        let cell = UITableViewCell(style: .subtitle, reuseIdentifier: nil)
+        switch properties[indexPath.row] {
+        case .name:
+            cell.textLabel?.text = "Name"
+            cell.detailTextLabel?.text = name ?? currentUserController.currentUser?.name
+            let button = UIButton(type: .detailDisclosure, primaryAction: UIAction(handler: { _ in
+                self.presentAlert(title: "Name", textFieldPlaceholder: self.currentUserController.currentUser?.name) { newValue in
+                    self.name = newValue
+                    self.updateUserData()
+                }
+            }))
+            button.setImage(.init(systemName: "pencil"), for: .normal)
+            cell.accessoryView = button
+        case .readReceiptsEnabled:
+            cell.textLabel?.text = "Read Receipts Enabled"
+            cell.accessoryView = makeSwitchButton(UserConfig.shared.readReceiptsEnabled ?? true) { newValue in
+                UserConfig.shared.readReceiptsEnabled = newValue
+            }
+        case .typingIndicatorsEnabled:
+            cell.textLabel?.text = "Typing Indicators Enabled"
+            cell.accessoryView = makeSwitchButton(UserConfig.shared.typingIndicatorsEnabled ?? true) { newValue in
+                UserConfig.shared.typingIndicatorsEnabled = newValue
+            }
         }
+        return cell
+    }
+
+    private func synchronizeAndUpdateData() {
+        currentUserController.synchronize()
+        updateUserData()
     }
 
     private func updateUserData() {
-        nameTextField.text = currentUserController.currentUser?.name ?? "Unknown"
-
         guard let imageURL = currentUserController.currentUser?.imageURL else { return }
-
         DispatchQueue.global().async { [weak self] in
             guard let data = try? Data(contentsOf: imageURL), let image = UIImage(data: data) else { return }
             DispatchQueue.main.async {
                 self?.imageView.image = image
             }
         }
+
+        if let typingIndicatorsEnabled = currentUserController.currentUser?.privacySettings.typingIndicators?.enabled {
+            UserConfig.shared.typingIndicatorsEnabled = typingIndicatorsEnabled
+        }
+        if let readReceiptsEnabled = currentUserController.currentUser?.privacySettings.readReceipts?.enabled {
+            UserConfig.shared.readReceiptsEnabled = readReceiptsEnabled
+        }
+
+        tableView.reloadData()
     }
 
     @objc private func didTapUpdateButton() {
-        guard let newName = nameTextField.text, newName != currentUserController.currentUser?.name else { return }
-        currentUserController.updateUserData(name: newName) { [weak self] _ in
-            self?.synchronizeAndUpdateData()
-        }
+        currentUserController.updateUserData(
+            name: name,
+            privacySettings: .init(
+                typingIndicators: UserConfig.shared.typingIndicatorsEnabled.map { .init(enabled: $0) },
+                readReceipts: UserConfig.shared.readReceiptsEnabled.map { .init(enabled: $0) }
+            )
+        )
     }
 
     func currentUserController(_ controller: CurrentChatUserController, didChangeCurrentUser: EntityChange<CurrentChatUser>) {
+        name = controller.currentUser?.name
         updateUserData()
+    }
+
+    private func makeSwitchButton(_ initialValue: Bool, _ didChangeValue: @escaping (Bool) -> Void) -> SwitchButton {
+        let switchButton = SwitchButton()
+        switchButton.isOn = initialValue
+        switchButton.didChangeValue = didChangeValue
+        return switchButton
     }
 }