+++ Sources/StreamChat/Controllers/ChannelController/ChannelController.swift
@@ -137,6 +137,16 @@ public class ChatChannelController: DataController, DelegateCallable, DataStoreP
 
     private var markingRead: Bool = false
 
+    /// A boolean value indicating if it should send typing events.
+    /// It is `true` if the channel typing events are enabled as well as the user privacy settings.
+    internal var shouldSendTypingEvents: Bool {
+        /// Ignore if user typing indicators privacy settings are disabled. By default, they are enabled.
+        let currentUserPrivacySettings = client.currentUserController().currentUser?.privacySettings
+        let isTypingIndicatorsForCurrentUserEnabled = currentUserPrivacySettings?.typingIndicators?.enabled ?? true
+        let isChannelTypingEventsEnabled = channel?.canSendTypingEvents ?? true
+        return isTypingIndicatorsForCurrentUserEnabled && isChannelTypingEventsEnabled
+    }
+
     /// Set the delegate of `ChannelController` to observe the changes in the system.
     public var delegate: ChatChannelControllerDelegate? {
         get { multicastDelegate.mainDelegate }
@@ -600,8 +610,8 @@ public class ChatChannelController: DataController, DelegateCallable, DataStoreP
     /// - Parameter completion: a completion block with an error if the request was failed.
     ///
     public func sendKeystrokeEvent(parentMessageId: MessageId? = nil, completion: ((Error?) -> Void)? = nil) {
-        /// Ignore if typing events are not enabled
-        guard channel?.canSendTypingEvents == true else {
+        /// Ignore if app-level typing events or user-level typing events are not enabled.
+        guard shouldSendTypingEvents else {
             callback {
                 completion?(nil)
             }
@@ -630,8 +640,8 @@ public class ChatChannelController: DataController, DelegateCallable, DataStoreP
     /// - Parameter completion: a completion block with an error if the request was failed.
     ///
     public func sendStartTypingEvent(parentMessageId: MessageId? = nil, completion: ((Error?) -> Void)? = nil) {
-        /// Ignore if typing events are not enabled
-        guard channel?.canSendTypingEvents == true else {
+        /// Ignore if app-level typing events or user-level typing events are not enabled.
+        guard shouldSendTypingEvents else {
             channelFeatureDisabled(feature: "typing events", completion: completion)
             return
         }
@@ -658,8 +668,8 @@ public class ChatChannelController: DataController, DelegateCallable, DataStoreP
     /// - Parameter completion: a completion block with an error if the request was failed.
     ///
     public func sendStopTypingEvent(parentMessageId: MessageId? = nil, completion: ((Error?) -> Void)? = nil) {
-        /// Ignore if typing events are not enabled
-        guard channel?.canSendTypingEvents == true else {
+        /// Ignore if app-level typing events or user-level typing events are not enabled.
+        guard shouldSendTypingEvents else {
             channelFeatureDisabled(feature: "typing events", completion: completion)
             return
         }