+++ CHANGELOG.md
@@ -5,6 +5,22 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ### 🔄 Changed
 
+# [4.53.0](https://github.com/GetStream/stream-chat-swift/releases/tag/4.53.0)
+_April 30, 2024_
+
+## StreamChat
+### ✅ Added
+- Add user privacy settings to control user events [#3142](https://github.com/GetStream/stream-chat-swift/pull/3142)
+- Add `privacySettings` parameter to `connectUser()` and `CurrentUserController.updateUserData()` [#3142](https://github.com/GetStream/stream-chat-swift/pull/3142)
+- Expose `ChatCurrentUser.privacySettings` [#3142](https://github.com/GetStream/stream-chat-swift/pull/3142)
+- Adds `reactionGroups` to `ChatMessage` so that reactions can be sorted by date in the message list [#3158](https://github.com/GetStream/stream-chat-swift/pull/3158)
+
+## StreamChatUI
+### ✅ Added
+- Adds `firstReactionAt` and `lastReactionAt` to `ChatMessageReactionData` to be able to sort by dates in `Components.reactionsSorting` [#3158](https://github.com/GetStream/stream-chat-swift/pull/3158)
+- Add `ChatReactionListController` to query and filter reactions from a message [#3167](https://github.com/GetStream/stream-chat-swift/pull/3167)
+- Add `ChatClient.reactionListController()`  to create an instance of `ChatReactionListController` [#3167](https://github.com/GetStream/stream-chat-swift/pull/3167)
+
 # [4.52.0](https://github.com/GetStream/stream-chat-swift/releases/tag/4.52.0)
 _April 09, 2024_